pyproject.toml
[project]
name = "my-website-check"
version = "0.1.0"
description = "AWS Lambda function to monitor website for broken images"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/website_check"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
